## Summary
This is the **aggregate-only** split from the all-in-one managed API
work in #7154. It builds on the scalar callback foundation and
extension-loading work that already landed, and intentionally leaves
custom collations and SDK/.NET exposure for separate PRs.
This PR adds managed aggregate callback support in core by:
- extending aggregate registration/discovery to carry callback context,
value cleanup, aggregate-state cleanup, and arity metadata through the
extension API
- wiring `AggFunc::External` through aggregate planning/translation for
fixed and variadic call-site arity
- extending VDBE aggregate step/finalization paths for external
aggregate state, grouped and ungrouped aggregation, zero-row
finalization, and callback error cleanup
- adding integration coverage for grouped/ungrouped aggregates,
`DISTINCT`, `FILTER`, variadic arity, zero-row finalization, and
step/finalizer error recovery
## Related split PRs
- Source all-in-one PR: #7154
- Scalar callback foundation: #7205 (merged)
- Extension loading foundation: #7206 (merged)
- Custom collations: #7207 (independent / out of scope)
- This PR: managed aggregate callbacks only
## Explicitly out of scope
- SDK-kit / .NET exposure
- custom collations (#7207)
- extension loading implementation (#7206)
- scalar callback foundation (#7205)
- workflows, simulator changes, and unrelated source-branch hunks
## Validation
- `cargo fmt --check`
- `git diff --check`
- `cargo check -q -p core_tester --features pure-rust-crypto --test
integration_tests`
- `cargo clippy -q -p core_tester --features pure-rust-crypto --test
integration_tests -- --deny=warnings`
- `cargo test -q -p core_tester --features pure-rust-crypto --test
integration_tests managed_aggregate -- --nocapture`


Closes #7271

+++ core/connection.rs
@@ -2920,7 +2920,7 @@ impl Connection {
             .map(|f| {
                 let is_agg = f.func.is_aggregate();
                 let argc = match &f.func {
-                    function::ExtFunc::Aggregate { argc, .. } => *argc as i32,
+                    function::ExtFunc::Aggregate { argc, .. } => *argc,
                     function::ExtFunc::Scalar { argc, .. } => *argc,
                 };
                 (
